add default value for 'cur_pod'
JIRA: YARDSTICK-551

As a Yardstick user, I want to use 'task_args' in the test suite yaml to
overwrite the pre-defined arguments in the test case yaml. Current framework
requires a 'NODE_NAME' when searching for 'task_args'.
A default value should be given when the
'NODE_NAME' is undefined in the environment.

Change-Id: I4e0ed2c982632c27d8bdcb6db7230ebe38201abb
Signed-off-by: JingLu5 <lvjing5@huawei.com>