ci in huawei-pod1 do not run test case after tc074
JIRA: YARDSTICK-568

https://build.opnfv.org/ci/view/yardstick/job/yardstick-compass-baremetal-daily-master/814/consoleFull
the test case tc075 and tc027 after tc074 does not run in the ci, which
is not correct.
the logs shows the suite quit after tc074 finishes.

The reason is:
In this patch: https://gerrit.opnfv.org/gerrit/#/c/26763/
I add a task_id to the context name, but in tc074, it do not have a
name.
It will cause a keyError.

Change-Id: I09608675d9dadcc3ff0d825ca63db3be5e414120
Signed-off-by: chenjiankun <chenjiankun1@huawei.com>