heat: replace dict key checking with .get()
We init all the fields to None, so the default value
is None, which is what .get() returns when the key
is not present it the dictionary.

Replace

  if key in dict:
    self.val = dict[key]

with

  self.val = dict.get(key)

This also has the added beneifit of re-initializing
default values to None.

In idiomatic Python we do not check a dictionary twice
if we can avoid it.  Either use dict.get() with a default
value or catch the KeyError and do the correct thing.

Change-Id: If8d1caeb7288f888f02622d62e8db9f59f50717a
Signed-off-by: Ross Brattain <ross.b.brattain@intel.com>