pathspec 'master' did not match any file(s) known to git
JIRA: YARDSTICK-605

On stable branch, exec_tests.sh will also try to git checkout master
branch, which cause error "pathspec 'master' did not match any file(s)
known to git." This is a remianing bug from the previous release.
see at:
https://build.opnfv.org/ci/view/yardstick/job/yardstick-apex-lf-pod1-daily-danube/62/console
This error occur in line 48.
But the reason is 'set -e'
If we set 'set -e':
'git checkout master && git pull' return 1 but it will go on executing;
but 'git checkout master' will return 1 and terminate immediately.
If we do not set, it will return 1 but will terminate immediately.
Actually we not need 'git checkout master' here because it is already in
stable/danube branch.
So I replace it with '$YARDSTICK_BRANCH'.

Change-Id: I106a1da28cf2deee90ebcb25adaf638a210928ee
Signed-off-by: chenjiankun <chenjiankun1@huawei.com>