yardstick offline support
JIRA: YARDSTICK-597

Currently yardstick can not run offline.
In some case this will block the process, causing yardstick job failed
The reason is yardstick will pull the latest code when run in CI.
Actually we need not pull the latest code because the code in yardstick
is already the latest code.
So I remove them.
Also I check the environment variable and yardstick-image when fetch
openrc file and load images.

Change-Id: I797da272932ab5292aa40585c82c119aa71b8849
Signed-off-by: chenjiankun <chenjiankun1@huawei.com>