Add SRIOV support
A generic provider network solution is introduced, where any existing provider network can be used. To identify whether a network specified in the test case is a provider network new attributes are introduced in the test case network section:
networks:
test-net1:
cidr: '192.168.1.0/24'
provider: sriov
net: sriov-net1
subnet: sriov-subnet1
test-net2:
cidr: '192.168.2.0/24'
provider: sriov
net: sriov-net2
subnet: sriov-subnet2
If the provider attribute is present, the network is an existing provider network. If the value is sriov, binding:vnic_type=direct is added to the interface in the heat deployment template. In orchestrator/heat.py, the interface creating functions are given a new parameter that tells if the network in use is a provider network. The benchmark/contexts/model.py is changed to store the value of the provider attribute from the test case and function calls to port creation is updated with the provider parameter. The same change is made in contexts/heat.py as well. Also calls for creating a new (tenant) network was surpressed if the provider attribute is present.

Add: test_model.py

Change-Id: I76004c4fcc9bffcfd8ed021fd647e0cecb346ef4
JIRA: YARDSTICK-612
Signed-off-by: Jing Zhang <jing.c.zhang@nokia.com>