Enable NFVi collectd to properly collect KPIs during test run
JIRA: YARDSTICK-679

Some of the collectd NFVi KPIs was missing during collect_kpi call due
to aync behavior of the collectd. Now we move the collection to a
separate process and cache it in queue and allow the KPIs to read out of
Queue during collect_kpi invocation. Also, terminate amqp client process
and stop collectd after test completion.

Change-Id: I9968ae571b6db7f350bcebae1ec40bddd30b2f73
Signed-off-by: Deepak S <deepak.s@linux.intel.com>