Update patch set 9
Patch Set 9:

Hello JingC,
I have a little question about the Method to realize SR-IOV?

1. Do we need to set up a new network and subnet manually.
2.If we need to bind the provider network with a way that the network and subnet will be created manually.

I mean when we need to deploy a SR-IOV environment, There is no doubt to have a physics network which have SR-IOV ability. but when we use this physics network to create a provider network is not different with the normal condition. 
And the "provider" this field may be used to create provider network(such like vlan vxlan flat). So I think it may be not good for us to bind this filed with creating network manually.
Could you please answer my question? Thank you very much.

Patch-set: 9