Update patch set 9
Patch Set 9:

To Ace:
1. Do we need to set up a new network and subnet manually.

[Jing] Yes, the referred provider network has to exist.

2.If we need to bind the provider network with a way that the
network and subnet will be created manually.I mean when we need to deploy a SR-IOV environment, There is no
doubt to have a physics network which have SR-IOV ability. but when
we use this physics network to create a provider network is not
different with the normal condition.
And the "provider" this field may be used to create provider
network(such like vlan vxlan flat). So I think it may be not good
for us to bind this filed with creating network manually.
Could you please answer my question? Thank you very much.
I mean : Could we use the heat-template to create the network "sriov-net1" and "sriov-net2" ?

[Jing] SRIOV support is different from OVS. This "provider" network based approach is a simple and straightforward, it can be used to test both SRIOV and OVS. For sure there is other way to do this, but I trust you will be looking at much bigger development effort.

Patch-set: 9