Update patch set 9
Patch Set 9:

Hi Ace,

I am not sure in which way I can help? This code change is very small, maybe you can clone it, give it a try, and see if you like it?

We use yardstick to benchmark our stack product, in order to do the work, we have done code extensions like this one. They are the type of fast small extensions out of need, we did not plan them, they are developed "on the go".

Patch-set: 9