constants: default SERVER_IP to 127.0.0.1
When running unittest while connected to P2P VPN,
their is no default route, so we get None.

Instead of using None, default to 127.0.0.1

Not sure why this was renamed to SERVER_IP either,
what do we use this value for?

Still getting py.test unittest failures all the time.

The problem is we lookup influxdb.ip but SERVER_IP is None,
so the lookup fails.

  INFLUXDB_IP = get_param('influxdb.ip', SERVER_IP)

SERVER_IP can't ever be None


Change-Id: Iedb09dc541137f7cdc6ef8d26dd312807985bfa8
Signed-off-by: Ross Brattain <ross.b.brattain@intel.com>