Update patch set 9
Patch Set 9:

May be the simple solution for concurrency is to use two ports, the first port is an OVS port for yardstick control, the 2nd port is the SRIOV port for fast data path.

This is a simple solution. I can add it quickly. I will wait for your comment first.

Patch-set: 9
Label: Verified=0