Update patch set 2
Patch Set 2:

> Jack, what do we use SERVER_IP for, and why is it the gateway?

We use docker in default. For example, we will use the ip to visit influxdb service. Since influxdb docker container is out of our yardstick docker container. And we do not know the ip of docker container. But we can get the gateway ip. If we send request to the gateway ip, it will pass to the host server. Then the request will pass to the influxdb container because of the port mapping.

Patch-set: 2
Label: Verified=0