Update patch set 2
Patch Set 2: Code-Review-1

A curios question. Why do we have the disable_trace.sh on yardstick? IMHO, the enabling/disabling trace should not be part of the benchmark, although it will be  good if we can have a mechanism provided by yardstick that we can provide post-scripts.

Patch-set: 2
Label: Code-Review=-1