Update patch set 2
Patch Set 2:

Hi, Maciej,
For now, we run Yardstick from an x86 machine, although the target nodes are AArch64 (there is a related change adding AArch64 support as well - https://gerrit.opnfv.org/gerrit/#/c/34841/).
This will probably break for Armband, but if it's urgent to fix the x86 case, we can merge it like this, and implement it better as part of Alex's change linked above.

Patch-set: 2