Adjust MTU inside instance to avoid SSH hanging
In arm-pods we have mtu 9000 in the Openstack infrastructure.
However, the Yardstick docker and bridge use MTU 1500.
Due to fragmentation/reassembly issues along the path between the 
container and the instance, the SSH connection hangs when 
transferring output produced by ramspeed json format results.
To avoid this and keep the change local to these tests,
we change the mtu inside the instance during the tests.
It makes the test work with minimum changes.

JIRA: ARMBAND-271

Change-Id: If069aaeab4f738c55de6f878ee1f3a7941868196
Signed-off-by: Alexandru Nemes <alexandru.nemes@enea.com>