don't git pull when cloning in exec_tests.sh
why would we want to git pull?

We just checked out the version we need.

If we git pull we get an error

HEAD is now at abbf19fc... Merge "prototypes: xci: Add
'ansible-hardening' role to the requirements file"
You are not currently on a branch.
Please specify which branch you want to merge with.
See git-pull(1) for details.

    git pull <remote> <branch>

Change-Id: I8ea9b64e3a785f271bf13ecb2bdd22fe99655bfe
Signed-off-by: Ross Brattain <ross.b.brattain@intel.com>