Blacklist TC011 from suites with ODL
In deployments with ODL, security groups behave badly because of an
IPv6 rule that causes IPv4 packets to be incorrectly routed.
This is described in:
https://jira.opnfv.org/browse/ARMBAND-61
As a consequence, Yardstick TC011 fails because iperf cannot 
create a connection.
The proposed solution is to temporarily comment out TC011
from test suites with ODL.
As mentioned here, security groups support for IPv6 is partial:
artifacts.opnfv.org/ipv6/docs/gapanalysis/gap-analysis-odl-boron.html
On nosdn deployments this issue is not reproducing and the test runs ok.

JIRA: ARMBAND-269

Change-Id: Id5fede27262c21d777ec8e9988941643d16f9417
Signed-off-by: Alexandru Nemes <alexandru.nemes@enea.com>