yardstick env influxdb/grafana cmd support centos
JIRA: YARDSTICK-714

Currently yardstick env influxdb/grafana command do not support centos.
Because we use the gateway ip to get the service of influxdb and grafana.
But in centos, we can not access influxdb/grafana service via gateway ip.
In this patch, I use docker inspect to get the ip of influxdb and grafana.
So these command can support centos.

Change-Id: I4599cbbd0fe43de9cf08e5d9e74d31edbb742998
Signed-off-by: chenjiankun <chenjiankun1@huawei.com>