Update patch set 1
Patch Set 1:

> why stable/danube?

In that patch:
https://gerrit.opnfv.org/gerrit/#/c/37483/

I changed utils.parse_ini_file method, when the path is not exist, it will raise a
RuntimeError, but sadly this method called by task.py, so yardstick.conf is must. So I am thinking we can put this file to /etc/yardstick by default.

sorry for cause this bug!

Patch-set: 1