Bugfix: yardstick always report 'PASS' to DB
JIRA: YARDSTICK-768

Now yardstick will not stop if test case failed,
so that we can run the rest test case in the suite.

CI judge task status by the cmd return code. In this way,
the 'yardstick task start' cmd will always return 0.
So we will always report 'PASS' to DB.

In this patch I add a judgement in cmd, if the status is not 'PASS', we
will raise a RuntimeError.

Change-Id: I655424dd9cd3782869986963a17b24acfb340345
Signed-off-by: chenjiankun <chenjiankun1@huawei.com>