KVMFORNFV: Update Grafana config for Packet Forwarding test cases.
JIRA: YARDSTICK-783

This patch is used for updating the configuration file of opnfv grafana
dashboard for kvmfornfv packet forwarding to publish the results of new
test cases added as part of kvmfornfv daily job.

Also used for updating the config file of opnfv grafana dashboard for
kvmfornfv cyclictest test cases to correct the alias colours.

Change-Id: I34aa155e2782e58147e55f974134673a892eb93b
Signed-off-by: Gundarapu Kalyan Reddy <reddyx.gundarapu@intel.com>