Update patch set 4
Patch Set 4:

Yardstick-TC0XX-XXXXX is a backup for community grafana config, and that would make the config file under version control. I guess your new config will be used as a local config not for community. so that would be better if you keep the old config file instead of removing them.

Patch-set: 4
Label: Verified=0