ansible: resize VNF image
When compiling Trex, DPDK, collectd and all the SampleVNFs
we use more than the 2.2GB size of the original Ubuntu cloud image.

Accordingly we need to resize the image.

If we were not inside a docker container we would use virt-resize
to automatically handle all the cases, but virt-resize launches qemu.

Instead we can use qemu-img to add extra space, then
luckily we can use parted to resize the partition and finally
resize2fs to resize the filesystem.

This limits us to only ext3/4 images, but if we need to
we could add support for other filesystems by checking
file system type.

Change-Id: Iac84b8e6967af5be64c280a7b1eaaf09f5d6b3aa
Signed-off-by: Ross Brattain <ross.b.brattain@intel.com>