New common utils classes and methods for executing commands
Classes:
* ExecResultTuple
  A named tuple to collect the status, stdout and stderr from
the result of an executed command.  With helper methods to
check exit status and stderr for bad strings.

* ExecExecption
  An Exception class designed to handle building messages from
ExecResultTuple instances.

Methods:
* execute_shell_command
  A method to open a subprocess using check_output that
exposes all the arguments and keyword arguments but returns
an ExecResultTuple, putting the Exception traceback into the
stderr field should an Exception be raised.  Exit status is 0 when
no exception is caught and -1 when an exception is caught.
Several duplications of this method that were scattered through
the code base are now consolidated on this version.

Change-Id: I60ecf42932a56872805d805a5a3492a1084134d3
Signed-off-by: Edward MacGillivray <edward.s.macgillivray@intel.com>