Generate pod.yaml from current context
We want to generate pod.yaml from Heat contexts so we can
re-use the context without destroying it.

But we don't have node role information and it doesn't
make sense in this case, so make the role optional.

Since we changed Heat to use pkey instead of key_filename,
we can embed the pkey into the pod.yaml, but we have
to make sure to convert the pkey to string, in case
it is a RSAKey object

Change-Id: Ibefcfbd8236e68013a704c39964cb870da825da8
Signed-off-by: Ross Brattain <ross.b.brattain@intel.com>