prox: fix TG KPIs
The problem is that we share the same ProxResourceHelper
for both VNF and TG.

For VNF we want to talk to resource.py and get collectd KPIs.
For TG we need to read from the queue the TG calculated KPIs and
we also want collectd KPIs.

workaround is to use a different method name collect_collectd_kpi
for VNFs

Change-Id: Icc2132758e37ce210f5600a0cd433077930208e5
Signed-off-by: Ross Brattain <ross.b.brattain@intel.com>