Update patch set 1
Patch Set 1:

> looks good, but two questions
 > 1.As your patch will set the IP_Routing as default for UDP_Stream
 > test. I wonder if there is any performance change against
 > no_IP_routing when we run test in one subnet with IP_routing
 > setting. should we make IP_ROUTING as a configure parameter?
 > 
 > 2.  I guess, basically, netperf test case just run in one subnet
 > with existing test yaml, is that mean that you did some extension
 > for netperf test case? would you upload the test yaml as well?

Hi Kube:

Thanks for your question.
1. I checked with the src code of Netperf, for the UDP_STREAM, it just set the SO_DONTROUTE on the data socket for the reason:
"to cover the backsides of incompetent testers who have bogus setups"
So I think just enabling the original routing option on the data socket would not hurt the performance, which are also certified with my given links in the commit msg.

2. There are many cases that we would have host, target on different subnets. 
For example, for Kubernetes with Flannel CNI(which is the most common use case), each node would have a different subnet configured. So if the host,target Pod are scheduled to 2 different nodes(which is the usual case for k8s), the host and target will each have a different subnet. So we should enable the routing for netperf UDP_STREAM test, otherwise the test can not be executed at all.

Patch-set: 1