Get node IPs and IDs according to env
Extract node IPs and IDs for each node having a controller or compute 
role (name starting with the "cmp" or "ctl" prefix) and add them into 
the $pod_yaml file, analogous to the previous implementation.
Since node IDs are expected to be unique and integers (condition that 
does not match the salt node ID format), they are substituted with an 
incremented index for each controller/compute node in the environment.

Change-Id: Id90626edc3f098bd96343336b2be179721dee5a1
Signed-off-by: Catalina Focsa <catalina.focsa@enea.com>