Update patch set 21
Patch Set 21: Code-Review-1

using classes to hold methods that should be module-scope is not good

if we want to have a method called "check_if_vm_exists_and_delete" and that method is static or class (and not instance) and calls only other static or class methods, then it and the methods it needs should be module functions.

in libvirt.py:

 def check_if_vm_exists_and_delete(...)
   ...

in sriov.py:

 from libvirt import check_if_vm_exists_and_delete
 ...
 def my_func(...):
   my_value = check_if_vm_exists_and_delete(...)

modules are collections of static and class methods, modules are like classes but they don't normally have instance methods.

what are we getting for creating instances of Libvirt, i don't see any reason to have instances of the class.

Patch-set: 21
Label: Code-Review=-1