drain runner queue and undo cancel_join_thread
Sometimes the runners can hang.  Initially
debugging lead to the queue join thread, so I thought
we could cancel all the join threads and everything would be okay.

But it turns out canceling the queue join threads can lead
to corruption of the queues, so when we go to drain the queues
the task hangs.

But it also turns out that we were not properly draining
the queues in the task process.  We were waiting for all
the runners to exit, then draining the queues.

This is bad and will cause the queues to fill up and hang
and/or drop data or corrupt the queues.

The proper fix seems to be to draining the queues in a
loop before calling join with a timeout.

Also modified the queue drain loops to no block on queue.get()

Revert "cancel all queue join threads"

This reverts commit 75c0e3a54b8f6e8fd77c7d9d95decab830159929.

Revert "duration runner: add teardown and cancel all queue join threads"

This reverts commit 7eb6abb6931b24e085b139cc3500f4497cdde57d.

Change-Id: Ic4f8e814cf23615621c1250535967716b425ac18
Signed-off-by: Ross Brattain <ross.b.brattain@intel.com>