Update patch set 6
Patch Set 6:

the python code is fine, though i'd prefer to remove the class attribute which was added to allow UT of code that was "disabled".  i realize that this would entail more work to get the UTs as intended, but it was what i intended.

Patch-set: 6
Reviewer: Gerrit User 6831 <6831@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Verified=0