NSB: DPDK rebind drivers to detect MAC address
If no devices are present we can't detect MAC address so
we can't match Heat ports to interfaces.

We can use lspci to ask the kernel what driver it should use
for the PCI device.

We can then use dpkd-devind to find all the PCI address we care
about and create a map with PCI device and real kernel driver.

Then we can dpdk force rebind to the kernel driver.
Once we have rebound to the kernel driver we can detect
MAC address and all the other attributes that are required.

detect socket for interface

use numa_node to probe socket number, or guess
based on PCI BUS address >= 0x10

VnfSshHelper allow override of wait timeout

Change-Id: I14cb657ed289a77941d048345d06ced5b5d5da52
Signed-off-by: Ross Brattain <ross.b.brattain@intel.com>