task: drain background runner queues
We were not draining the queues in the background
runners.

Modify the main runner_join loop to loop over
all the runners and drain them.

The runner join method does extra work for periodic actions
so we can't change its behavior.

Instead create a new poll() method and use that
to check runner status

Change-Id: I9466ba40a6a4c45c82cedff279cbb4817c6b66ad
Signed-off-by: Ross Brattain <ross.b.brattain@intel.com>