ping: always save rtt data, influxdb ignore empty data
If the SLA was failing we were raising AssertionError and not storing
the rtt in the data dict.  This caused influxdb parse errors because the
data was empty.

Fixup influxdb to ignore records with no data, so we don't try to parse
no data.

Change the ping logic to always record the rtt result even if the SLA
was not met.

Also fixup ping logic in cases where ping does not return results.  If
SLA is defined use SLA * 10 otherwise use large float that doesn't break
the grafana scale too much, maybe 999999

JIRA: YARDSTICK-809

Change-Id: Id2d51216581644a80e8c7b9aa98919a766008adf
Signed-off-by: Ross Brattain <ross.b.brattain@intel.com>