resource: remove cores args
We used to try to probe all the cores that the VNF was using
and only dump CPU stats for those cores.

We can't really detect those core accurately and we would rather
dump all the core information and let influxdb and grafana filter 
the information.

We do end up with excessive KPI output, especially on systems with 88
cores, but this is manageable.

The core logic was partially removed, this finishes the removal.


Change-Id: I5cbb694fd982519e2df54db49a21ed5948e13537
Signed-off-by: Ross Brattain <ross.b.brattain@intel.com>