Fix test_scenario_list and improve functional tests
Test case test_scenario_show_Lmbench was not
capable of failing, since the literal string was always
"true".  Now these test cases define a substring and
that substring is asserted to be "in" the command
output.  This means that the substring and the
command output will be displayed upon failure,
instead of merely "False is not true".

Additionally, now we:
* use Popen instead of check_output
* use a string buffer instead of PIPE
* call super.__init__
* capture all input arguments as attributes
* don't define __del__ to do nothing
* don't catch exception to raise them
* catch exceptions to log them
* default __doc__ to None
* default __scenario_type__ to None

Change-Id: Id87f437d9ffac2e452fdece82de9875f1ab4a460
Signed-off-by: Edward MacGillivray <edward.s.macgillivray@intel.com>